add Custom Exceptions

+++ src/Core/HR.LeaveManagement.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using HR.LeaveManagement.Application.Contracts.Persistence;
+using HR.LeaveManagement.Application.Exceptions;
 using MediatR;
 
 namespace HR.LeaveManagement.Application.Features.LeaveType.Commands.DeleteLeaveType;
@@ -16,6 +17,11 @@ public class DeleteLeaveTypeCommandHandler : IRequestHandler<DeleteLeaveTypeComm
     public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
     {
         var leaveTypeToDelete = await _leaveTypeRepository.GetByIdAsync(request.Id);
+
+        if (leaveTypeToDelete == null)
+        {
+            throw new NotFoundException(nameof(Domain.LeaveType), request.Id);
+        }
 
         await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
